Isolated separate React applications
