+++ static/app/components/apps/Admin/PrivateRoute.js
@@ -28,7 +28,7 @@ class PrivateRoute extends React.Component {
 
 function mapStateToProps(state) {
   return {
-    authenticated: state.auth.authenticated
+    authenticated: state.admin.auth.authenticated
   };
 }
