+++ static/app/components/apps/Admin/auth/Login.js
@@ -68,8 +68,8 @@ class Login extends React.Component {
 
 function mapStateToProps(state) {
   return {
-    errorMessage: state.auth.error,
-    message: state.auth.message
+    errorMessage: state.admin.auth.error,
+    message: state.admin.auth.message
   };
 }
