+++ static/app/components/apps/Admin/pages/AdminsPage/actions/types.js
