+++ static/app/components/apps/Admin/pages/DashboardPage/reducers/Stats.js
