+++ static/app/reducers/apply.js
@@ -1,0 +1,3 @@
+export default {
+  // Insert any reducers relating to apply
+};
