+++ static/app/routes.js
@@ -1,14 +1,12 @@
 import React from 'react';
 import {Route, Switch} from 'react-router-dom';
 
-import Admin from './components/Admin';
-import Apply from './components/Apply';
-import NotFoundPage from './components/pages/NotFound';
+import Admin from './components/apps/Admin';
+import Apply from './components/apps/Apply';
 
 export default (
   <Switch>
     <Route path="/admin" component={Admin} />
     <Route path="/apply" component={Apply} />
-    <Route component={NotFoundPage} />
   </Switch>
 );
